python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
sentence-transformers==2.3.1
hdbscan==0.8.33
scikit-learn==1.4.0
numpy==1.26.3
pydantic==2.6.0
python-dotenv==1.0.0
pytest==8.4.2
pytest-asyncio==1.2.0
tenacity==8.2.3
pybreaker==1.0.1
redis==5.0.1
requests==2.31.0
supabase==2.9.1

# Multilingual support
langdetect==1.0.9
deep-translator==1.11.4
arabic-reshaper==3.0.0
python-bidi==0.4.2

# For variant matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# HTTP client for async requests
httpx==0.27.2
google-generativeai>=0.8.0

# Pinecone vector database
pinecone>=5.0.0

# OpenAI for embeddings
openai>=1.0.0
